+++ src/main/java/dev/isxander/controlify/platform/client/fabric/FAPIImplCreativeTabHelper.java
@@ -66,8 +66,7 @@ public class FAPIImplCreativeTabHelper implements CreativeTabHelper {
                 .toList();
     }
 
-    @Override
-    public int getPageForTab(CreativeModeTab tab) {
+    private int getPageForTab(CreativeModeTab tab) {
         if (FabricCreativeGuiComponents.COMMON_GROUPS.contains(tab)) {
             return getCurrentPage();
         }