+++ changelog.md
@@ -13,6 +13,8 @@ This version has the following targets:
 
 ## Changes
 
+- Fallback to default controller font set if the font mapping file does not exist for the controller's namespace
+- Add a controller icon for PSP (thanks \_cheburkot\_)
 - Add incomplete support for NeoForge 1.20.4 and 1.20.6
 - Fix compatibility with Fabric API 0.100+ on 1.20.6
 - Update to the SDL library, this will be redownloaded